+++ lib/ui/groups/GroupWidgets.dart
@@ -2819,21 +2819,29 @@ class _GroupPollCardState extends State<GroupPollCard>{
 
 class GroupMemberProfileImage extends StatefulWidget {
   final String? userId;
+  final GestureTapCallback? onTap;
 
-  GroupMemberProfileImage({this.userId});
+  GroupMemberProfileImage({this.userId, this.onTap});
 
   @override
   State<GroupMemberProfileImage> createState() => _GroupMemberProfileImageState();
 }
 
-class _GroupMemberProfileImageState extends State<GroupMemberProfileImage> {
+class _GroupMemberProfileImageState extends State<GroupMemberProfileImage> implements NotificationsListener {
   Uint8List? _imageBytes;
   bool _loading = false;
 
   @override
   void initState() {
     super.initState();
+    NotificationService().subscribe(this, Content.notifyUserProfilePictureChanged);
     _loadImage();
+  }
+
+  @override
+  void dispose() {
+    super.dispose();
+    NotificationService().unsubscribe(this);
   }
 
   @override
@@ -2843,7 +2851,7 @@ class _GroupMemberProfileImageState extends State<GroupMemberProfileImage> {
         : Image.asset('images/missing-photo-placeholder.png', excludeFromSemantics: true);
 
     return GestureDetector(
-        onTap: _onImageTap,
+        onTap: widget.onTap ?? _onImageTap,
         child: Stack(alignment: Alignment.center, children: [
           Container(
               decoration: BoxDecoration(shape: BoxShape.circle, image: DecorationImage(fit: BoxFit.cover, image: profileImage.image))),
@@ -2884,6 +2892,18 @@ class _GroupMemberProfileImageState extends State<GroupMemberProfileImage> {
       _loading = loading;
       if (mounted) {
         setState(() {});
+      }
+    }
+  }
+
+  // Notifications
+
+  @override
+  void onNotification(String name, param) {
+    if (name == Content.notifyUserProfilePictureChanged) {
+      // If it's current user - reload profile picture
+      if (widget.userId == Auth2().accountId) {
+        _loadImage();
       }
     }
   }